requirements.txt
acme==3.0.1
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohasupervisor==0.2.2b5
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiohttp-fast-zlib==0.2.0
aiooui==0.1.9
aiosignal==1.3.2
aiozoneinfo==0.2.1
anyio==4.8.0
astral==2.2
async-timeout==5.0.1
async_interrupt==1.2.0
atomicwrites-homeassistant==1.4.1
attrs==24.2.0
audioop-lts==0.2.1
awesomeversion==24.6.0
bcrypt==4.2.0
bleak==0.22.3
bleak-retry-connector==3.8.0
bluetooth-adapters==0.21.1
bluetooth-auto-recovery==1.4.2
bluetooth-data-tools==1.22.0
boto3==1.36.10
botocore==1.36.10
btsocket==0.3.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
ciso8601==2.3.2
cronsim==2.6
cryptography==44.0.0
distlib==0.3.9
envs==1.4
filelock==3.17.0
fnv-hash-fast==1.0.2
fnvhash==0.1.0
frozenlist==1.5.0
h11==0.14.0
habluetooth==3.14.0
hass-nabucasa==0.87.0
home-assistant-bluetooth==1.13.0
homeassistant==2025.1.4
httpcore==1.0.7
httpx==0.27.2
identify==2.6.6
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jmespath==1.0.1
josepy==1.15.0
lru-dict==1.3.0
MarkupSafe==3.0.2
mashumaro==3.15
multidict==6.1.0
nodeenv==1.9.1
orjson==3.10.12
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.1.0
propcache==0.2.1
psutil==6.1.1
psutil-home-assistant==0.0.1
pycares==4.5.0
pycognito==2024.5.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyRFC3339==2.0.1
PyRIC==0.1.6.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
securetar==2024.11.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
snitun==0.39.1
SQLAlchemy==2.0.36
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-telnetlib==3.13.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
uart-devices==0.1.0
ulid-transform==1.0.2
urllib3==1.26.20
usb-devices==0.4.5
uv==0.5.8
virtualenv==20.29.1
voluptuous==0.15.2
voluptuous-openapi==0.0.5
voluptuous-serialize==2.6.0
webrtc-models==0.3.0
winrt-runtime==2.3.0
winrt-Windows.Devices.Bluetooth==2.3.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.3.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.3.0
winrt-Windows.Devices.Enumeration==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Storage.Streams==2.3.0
yarl==1.18.3
